Fix tracked_jobs.score column type (integer -> numeric)
Found while testing the new analytics module against realistic data: the
column was created as integer, but the /6-dimension average scoring model
produces decimal totals (e.g. 3.8, 4.2). Saving any job scored under the
current model has been failing at the database level with "invalid input
syntax for type integer" ever since the rescale - confirmed live via the
actual /api/stamp4/simple-apply/tracker endpoint. The one real tracked job
predates the rescale and still holds an old integer-scale value (82),
which is why this had gone unnoticed - no job has been successfully saved
since. Verified the fix by saving a real decimal-score job through the
live endpoint after applying the migration.

+++ docs/supabase-schema.sql
@@ -5,7 +5,7 @@
   country text,
   location text,
   salary text,
-  score integer not null,
+  score numeric not null,
   decision text not null,
   status text not null default 'Saved',
   date_added timestamptz not null default now(),
@@ -24,6 +24,10 @@
 -- Run this against an existing database that predates the parsed_job column
 -- (added so "Why this score?" explanations work for already-saved jobs too):
 -- alter table tracked_jobs add column if not exists parsed_job jsonb;
+
+-- Run this against an existing database where score is still integer (the original type) - the
+-- /6-dimension average scoring model produces decimal totals, so saving fails without this:
+-- alter table tracked_jobs alter column score type numeric using score::numeric;
 
 create table custom_job_sources (
   id uuid primary key default gen_random_uuid(),
